initial commit for new biasparams framework, bor iv biasparams calculation, fix to squid sim code, simplifications of didv and fixes to didv plotting
