+++ qetpy/core/__init__.py
@@ -1,6 +1,7 @@
 from . import didv
 from .didv import *
 from ._iv import *
+from ._biasparams import *
 from ._ibis import *
 from ._noise import *
 from ._fitting import *
