+++ qetpy/core/didv/_base_didv.py
@@ -11,10 +11,6 @@ __all__ = [
     "compleximpedance",
     "complexadmittance",
     "squarewaveresponse",
-    "get_i0",
-    "get_ibias",
-    "get_tes_bias_parameters_dict",
-    "get_tes_bias_parameters_dict_infinite_loop_gain"
 ]
 
 
@@ -266,9 +262,8 @@ def complexadmittance(f, *, rsh=None, rp=None, r0=None, beta=None, l=None,
     return 1 / impedance
 
 
-def squarewaveresponse(t, sgamp, sgfreq, dutycycle=0.5, *, rsh=None, rp=None,
-                       r0=None, beta=None, l=None, L=None, tau0=None,
-                       gratio=None, tau3=None, **kwargs):
+def squarewaveresponse(t, sgamp, sgfreq, params, dutycycle=0.5,
+                       *, rsh=None, **kwargs):
     """
     Method for calculating the TES response to a square wave for a
     given model, depending on the parameters inputted (see Notes
@@ -283,37 +278,15 @@ def squarewaveresponse(t, sgamp, sgfreq, dutycycle=0.5, *, rsh=None, rp=None,
         The peak-to-peak size of the square wave jitter (in Amps)
     sgfreq : float
         The frequency of the square wave jitter (in Hz)
+    params : float
+        The fit parameters (A, B, ... tau1, tau2... ) of the model
+        to be generated
     dutycycle : float, optional
         The duty cycle of the square wave jitter (between 0 and 1).
         Default is 0.5.
     rsh : float, optional
         Shunt resistance of the TES circuit, unis of Ohms. Used by 1-,
         2-, and 3-pole models.
-    rp : float, optional
-        Parasitic resistance on the non-shunt side of the TES circuit,
-        units of Ohms. Used by 1-, 2-, and 3-pole models.
-    r0 : float, optional
-        The operating resistance of the TES, units of Ohms. Used by 2-
-        and 3-pole models.
-    beta : float, optional
-        The current sensitivity of the TES, defined as
-        beta = d(log R)/d(log I). Used by 2- and 3-pole models.
-    l : float, optional
-        The loop gain of the TES, defined as l = P0*alpha/(G*Tc). Used
-        by the 2- and 3-pole models.
-    L : float, optional
-        The inductance in the TES circuit, units of Henrys. Used by 1-,
-        2-, and 3-pole models.
-    tau0 : float, optional
-        The thermal time constant of the TES (in sseconds), defined as
-        tau0 = C/G. Used by 1-, 2-, and 3-pole models.
-    gratio : float, optional
-        The ratio of thermal conductances in the 3 thermal block models
-        that are used as the basis of the 3-pole model. Used in 3-pole
-        model only.
-    tau3 : float, optional
-        The extra pole  associated with the 3 thermal block model,
-        units of seconds. Used in 3-pole model only.
     kwargs : dict, optional
         Any extra keyword arguments passed are simply ignored.
 
@@ -323,563 +296,26 @@ def squarewaveresponse(t, sgamp, sgfreq, dutycycle=0.5, *, rsh=None, rp=None,
         The response of the TES to a square wave jitter, based on
         the inputted parameters.
 
-    Notes
-    -----
-    For the inputted values, there are three possible models to be
-    used:
-
-    1-pole model
-        - used if `rsh`,  `rp`, and `L` are passed
-        - has the form:
-            dV/dI = rsh + rp + 2.0j * pi * freq * L
-
-    2-pole model (see https://doi.org/10.1007/10933596_3)
-        - used if `rsh`,  `rp`, `r0`, `beta`, `l` and `L` are passed
-        - has the form:
-            dV/dI = rsh + rp + 2.0j * pi * freq * L + r0 * (1 + beta)
-                  + (r0 * (2 + beta) * l / (1 - l)
-                  / (1.0 + 2.0j * pi * freq * tau0 / (1 - l)))
-
-    3-pole model (see https://doi.org/10.1063/1.4759111)
-        - used if `rsh`,  `rp`, `r0`, `beta`, `l`, `L`, `gratio` and
-          `tau3` are passed
-        - has the form:
-            dV/dI = rsh + rp + 2.0j * pi * freq * L + r0 * (1 + beta)
-                  + r0 * (2 + beta) * l / (1 - l)
-                  / (1.0 + 2.0j * pi * freq * tau0 / (1 - l)
-                  - gratio / (1 - l) / (1.0 + 2.0j * pi * freq * tau3))
-
-    """
-
-    passed_args = locals()
-    poles = _pole_extractor(passed_args)
-
-    if poles == 1:
-        A, tau2, _  = _BaseDIDV._convertfromtesvalues(
-            (rsh, rp, L, 0),
-        )
-        return _BaseDIDV._convolvedidv(
-            t, A, 0, 0, 0, tau2, 0, sgamp, rsh, sgfreq, dutycycle,
-        )
-    if poles == 2:
-        A, B, tau1, tau2, _  = _BaseDIDV._convertfromtesvalues(
-            (rsh, rp, r0, beta, l, L, tau0, 0),
-        )
-        return _BaseDIDV._convolvedidv(
-            t, A, B, 0, tau1, tau2, 0, sgamp, rsh, sgfreq, dutycycle,
-        )
-    if poles == 3:
-        A, B, C, tau1, tau2, tau3, _  = _BaseDIDV._convertfromtesvalues(
-            (rsh, rp, r0, beta, l, L, tau0, gratio, tau3, 0),
-        )
-        return _BaseDIDV._convolvedidv(
-            t, A, B, C, tau1, tau2, tau3, sgamp, rsh, sgfreq, dutycycle,
-        )
-
-
-
-def get_i0(offset, offset_err, offset_dict, output_offset=None,
-           closed_loop_norm=None, output_gain=1,
-           lgc_invert_offset=False,
-           lgc_calibration_on=False, calibration_dict=None, 
-           lgc_diagnostics=False):
-    """
-    Gets and returns the current and uncertainty in the current
-    through the TES from a didv_fitresult dictonary (which
-    includes a parameter for the offset of the didv and the offset)
-    
-    Parameters
-    ----------
-    offset: float
-        The current offset of the dIdV as fit (i.e. as measured, without
-        any corrections)
-        
-    offset_err: float
-        The uncertainty in the current offset of the dIdV fit
-        
-    offset_dict: dict
-        Dictionary of offsets gotten from the IV sweep.
-        
-    output_offset: float, volts
-        The output offset gotten from the event metadata. In units of volts,
-        we correct for volts to amps conversion with the closed loop norm.
-        
-    closed_loop_norm: float, volts/amp=ohms
-        The constant from the metadata used to translate the voltage measured by
-        the DAQ into a current coming into the input coil of the SQUIDs. In units of
-        volts/amp = ohms.
-        
-    output_gain: float, dimensionless
-        The dimensionless gain for the front end electronics. Used to translate the
-        output_offset in units of volts to the equivilant value read in the DAQ in
-        units of volts.
-        
-    lgc_calibration_on : bool, optional
-        By default False (i.e. not using the calibration). If True, uses the calibration_dict
-        to more closely approximate how changing the output_offset changes the current
-        measured.
-        
-    calibration_dict : dict, optional
-        A dictonary of data used to more closely model the relationship between the
-        output_offset and the change in the measured current in the device. 
-        
-    lgc_diagnostics : bool, optional
-        Used if you want to see the raw currents and offsets and how they're
-        added together. Prints these out
-        
-    Returns
-    -------
-    i0 : float
-        The calculated absolute current through the TES
-    i0_err : float
-        The uncertainty in the absolute current through the TES
-        
-    """
-
-    # SQUID controller variable offset
-    delta_i_variable = 0
-    if output_offset is not None:
-
-        # check if "closed_loop_norm" available
-        if closed_loop_norm is None:
-            raise ValueError(
-                'ERROR: "closed_loop_norm" parameter is '
-                'required!'
-            )
-        
-        # IV sweep "i0_variable_offset" (check old names for back compatibility)
-        i0_variable_offset_sweep = None
-        if 'i0_variable_offset' in offset_dict.keys():
-            i0_variable_offset_sweep = offset_dict['i0_variable_offset']
-        elif ('i0_changable_offset_cal' in offset_dict.keys()) and (lgc_calibration_on is True):
-            i0_variable_offset_sweep  =  offset_dict['i0_changable_offset_cal']
-        elif ('i0_changable_offset_uncal' in offset_dict.keys()) and (lgc_calibration_on is False):
-            i0_variable_offset_sweep  =  offset_dict['i0_changable_offset_uncal']
-        else:
-            raise ValueError('ERROR: i0 variable offset not found in '
-                             '"ivsweep_result" dictionary!')
-
-        # current IV variable offset
-        if lgc_calibration_on is False:
-            i0_variable_offset = output_offset * output_gain/closed_loop_norm
-        else:
-            if calibration_dict is None:
-                raise ValueError('ERROR: must include calibration_dict if '
-                                 'lgc_calibration_on is True (i.e. being used)!'
-                                 'Check offset dicts')
-                                 
-            elif (output_gain != 50):
-                raise ValueError('ERROR: calibration only done for a gain of 50')
-                
-            elif (calibration_dict['model'] == 'twopartlinear'):
-                m1, m2, b1, b2 = calibration_dict['params']
-                if output_offset > 0.0:
-                    i0_variable_offset = output_offset * m1 + b1
-                else:
-                    i0_variable_offset = output_offset * m2 + b2
-                    
-            elif (calibration_dict['model'] == 'lookup_extrapolated'):
-                offsets_arr = calibration_dict['params']['lookup_x']
-                currents_arr = calibration_dict['params']['lookup_y']
-                params_low = calibration_dict['params']['params_low']
-                params_high = calibration_dict['params']['params_high']
-                
-                def linear(x, m, b):
-                    return x*m + b
-                
-                if output_offset in offsets_arr:
-                    i0_variable_offset = currents_arr[list(offsets_arr).index(offset)]
-                elif output_offset > 0.0:
-                    i0_variable_offset = linear(output_offset, *params_high)
-                else:
-                    i0_variable_offset = linear(output_offset, *params_low)
-                    
-            else:
-                raise ValueError('ERROR: unknown calibration_dict model')
-            
-       
-        # delta variable offset
-        delta_i_variable = i0_variable_offset - i0_variable_offset_sweep
-       
-        #alerts user when the current offset changed
-        if np.abs(delta_i_variable) > 1e-9:
-            print(" ")
-            print("----------------------------------------------------------------")
-            print("ALERT: Variable output voltage offset has changed since the IV sweep.")
-            print("Thus, it needs to be taken into account when calculating true I0!")
-            print("IV sweep variable offset [muAmps]: " + str(i0_variable_offset_sweep*1e6))
-            print("This dIdV dataset variable offset [muAmps]: " + str(i0_variable_offset*1e6))
-            print("RESULTS MAY NOT BE ACCURATE IF GAIN/OFFSET NON LINEARITIES!")
-            print("----------------------------------------------------------------")
-            print(" ")
-
-    # calculate new i0
-    current_didv = offset - delta_i_variable 
-    if lgc_invert_offset:
-        current_didv = -current_didv
-    current_err_didv = offset_err
-    i0 = current_didv - offset_dict['i0_off']
-    
-    if lgc_diagnostics:
-        print("Current as measured from dIdV: " + str(offset) + " amps")
-        print("Variable current as meaured from metadata for this dIdV: " + str(i0_variable_offset) + " amps")
-        print(" ")
-        print("Variable current as measured during IV when offsets were measured: " + 
-              str(i0_variable_offset_sweep) + " amps")
-        print("Current offset as measured from IV: " + str(offset_dict['i0_off']) + " amps")
-        print(" ")
-        print("Delta variable current: " + str(i0_variable_offset) + " - " + 
-             str(i0_variable_offset_sweep) + " = \n " + str(delta_i_variable) + " amps")
-        print(" ")
-        print("True current through TES: " + str(offset) + " - " + str(delta_i_variable) + " - \n" + 
-             str(offset_dict['i0_off']) + " = " + str(i0) + " amps")
-    
-    i0_err = np.sqrt((current_err_didv)**2.0 + (offset_dict['i0_off_err'])**2.0)
-    
-    if lgc_diagnostics:
-        print(" ")
-        print(" --------- ")
-        print(" ")
-        print("Current uncertainty from dIdV: " + str(current_err_didv) + " amps")
-        print("Current offset uncertainty from IV: " + str(offset_dict['i0_off_err']) + " amps")
-        print(" ")
-        print("Total current uncetainty: (("  + str(current_err_didv) + ")**2.0 + (" +
-              str(offset_dict['i0_off_err']) + ")**2.0 )**0.5 = \n" + 
-              str(i0_err) + " amps")
-    
-    return i0, i0_err
-
-def get_ibias(tes_bias, offset_dict, lgc_diagnostics=False):
-    """
-    Gets and returns the current and uncertainty in the current
-    used to bias the TES from a metadata list
-    
-    Parameters
-    ----------
-        
-    tes_bias: float
-        The ibias gotten from the event metadata, i.e. without correcting for
-        the ibias offset calculated from the IV curve
-        
-    offset_dict: dict
-        Dictionary of offsets gotten from the IV sweep.
-        
-    lgc_diagnostics : bool, optional
-        Used if you want to see the raw currents and offsets and how they're
-        added together. Prints these out
-        
-    Returns
-    -------
-    ibias : float
-        The calculated absolute current used to bias the TES
-    ibias_err : float
-        The uncertainty in the absolute current used to bias the TES
-        
     """
     
-    ibias = tes_bias - offset_dict['ibias_off']
-    ibias_err = offset_dict['ibias_off_err']
-    
-    if lgc_diagnostics:
-        print("Bias current from metadata: " + str(tes_bias) + " amps")
-        print("Bias current offset from IV: " + str(offset_dict['ibias_off']) + " amps")
-        print("True bias current: " + str(tes_bias) + " - " + str(offset_dict['ibias_off']) + 
-             " = \n" + str(ibias) + " amps")
-        print(" ")
-        
-        print("Bias current uncertainty from IV: " + str(offset_dict['ibias_off_err']) + " amps")
-        
-    return np.absolute(ibias), ibias_err
+    if rsh is None:
+        rsh = 5e-3
 
-
-def _get_v0(i0, i0_err, ibias, ibias_err, rsh, rp):
-    """
-    Gets and returns voltage and uncertainty in voltage across the
-    TES at a given bias point.
+    A = params['A']
+    B = params['B']
+    C = params['C']
+    tau1 = params['tau1']
+    tau2 = params['tau2']
+    tau3 = params['tau3']
+    dt = params['dt']
     
-    Parameters
-    ----------
-        
-    i0 : float
-        The current through the TES at a given bias point
-        
-    i0_err: float
-        The uncertainty in the current through the TES at a given
-        bias point
-        
-    ibias: float
-        The bias current applied to the TES/shunt system
-        
-    ibias_err: float
-        The uncertainty in the bias current applied to the TES/shunt
-        system
-        
-    rsh: float
-        The shunt resistance in ohms
-        
-    rp: float
-        The parasitic resistance in ohms
-        
-    Returns
-    -------
-    v0 : float
-        The calculated voltage across the TES in volts
-        
-    v0_err : float
-        The calculated uncertainty in the voltage across the TES
-        in volts
-        
-    """
-    
-    ibias = np.absolute(ibias)
-    i0 = np.absolute(i0)
-    vb = rsh * (ibias - i0) #voltage across the shunt resistor
-    vb_err = rsh * np.sqrt(i0_err**2 + ibias_err**2)
-    
-    v0 = np.absolute(vb - rp * i0) #the voltage across the shunt resistor minus
-                      #the voltage across the parasitic resistance
-    v0_err = np.sqrt(vb_err**2 + rp**2 * i0_err**2)
-    
-    return v0, v0_err
-
-def _get_r0(i0, i0_err, v0, v0_err):
-    """
-    Gets and returns the resistance and uncertainty in resistance
-    of a TES at a given bias point
-    
-    Parameters
-    ----------
-        
-    i0 : float
-        The current through the TES at a given bias point
-        
-    i0_err: float
-        The uncertainty in the current through the TES at a given
-        bias point
-        
-    v0: float
-        The voltage across the TES at the bias point
-        
-    v0_err: float
-        The uncertainty in the voltage across the TES at the bias
-        point
-        
-    Returns
-    -------
-    r0 : float
-        The calculated resistance of the TES at the bias point
-        in ohms
-        
-    r0_err : float
-        The calculated uncertainty in the resistance of the TES at
-        the bias point in ohms
-        
-    """
-    
-    r0 = np.absolute(v0/i0)
-    r0_err = np.sqrt(v0_err**2 * i0**-2 + v0**2 * i0_err**2 * i0**-4)
-    
-    return r0, r0_err
-
-def _get_p0(i0, i0_err, v0, v0_err):
-    """
-    Gets and returns the bias power and uncertainty in bias power
-    of a TES at a given bias point
-    
-    Parameters
-    ----------
-        
-    i0 : float
-        The current through the TES at a given bias point
-        
-    i0_err: float
-        The uncertainty in the current through the TES at a given
-        bias point
-        
-    v0: float
-        The voltage across the TES at the bias point
-        
-    v0_err: float
-        The uncertainty in the voltage across the TES at the bias
-        point
-        
-    Returns
-    -------
-    p0 : float
-        The calculated bias power of the TES at the bias point
-        in watts
-        
-    p0_err : float
-        The calculated uncertainty in the bias power of the TES at
-        the bias point in watts
-        
-    """
-    
-    p0 = np.absolute(v0*i0)
-    p0_err = np.sqrt(v0_err**2 * i0**2 + v0**2 * i0_err**2)
-    
-    return p0, p0_err
-
-def get_tes_bias_parameters_dict(i0, i0_err, ibias, ibias_err, rsh, rp,
-                                 rn=None):
-    """
-    Gets and returns a dictonary of i0, v0, r0, and p0 with uncertainties
-    from the measured TES bias current and i0
-    
-    Parameters
-    ----------
-        
-    i0 : float
-        The current through the TES at a given bias point
-        
-    i0_err: float
-        The uncertainty in the current through the TES at a given
-        bias point
-        
-    ibias: float
-        The bias current applied to the TES/shunt system
-        
-    ibias_err: float
-        The uncertainty in the bias current applied to the TES/shunt
-        system
-        
-    rsh: float
-        The shunt resistance in ohms
-        
-    rp: float
-        The parasitic resistance in ohms
-
-    rn: float (optional) 
-        The normal resistance in ohms
-        
-    Returns
-    -------
-    bias_parameters_dict : dict
-        A dictonary of bias parameters and uncertainties, including i0, r0,
-        v0, and p0
-        
-    """
-    
-    i0 = np.absolute(i0)
-    
-    rl = rp + rsh
-    
-    v0, v0_err = _get_v0(i0, i0_err, ibias, ibias_err, rsh, rp)
-    r0, r0_err = _get_r0(i0, i0_err, v0, v0_err)
-    p0, p0_err = _get_p0(i0, i0_err, v0, v0_err)
-
-    if rn is None:
-        rn = np.nan
-
-    
-    bias_parameter_dict = {
-        'i0': i0,
-        'i0_err': i0_err,
-        'v0': v0,
-        'v0_err': v0_err,
-        'r0': r0,
-        'r0_err': r0_err,
-        'p0': p0,
-        'p0_err': p0_err,
-        'rp': rp,
-        'rsh': rsh,
-        'rshunt': rsh,
-        'rl': rl,
-        'ibias': ibias,
-        'rn': rn,
-    }
-    
-    return bias_parameter_dict
-    
-def get_tes_bias_parameters_dict_infinite_loop_gain(params, cov,
-                                                    ibias, ibias_err,
-                                                    rsh, rp, rn=None):
-    """
-    Gets and returns a dictonary of i0, v0, r0, and p0 with uncertainties
-    using the infinte loop gain approximation
-    
-    Parameters
-    ----------
-  
-    params : dict
-        The parameters (A, B, tau_1, etc.) of the previous dIdV fit.
-               
-    cov : matrix
-        The covariance matrix for the params, starting with the A, B
-        components.
-        
-    ibias: float
-        The bias current applied to the TES/shunt system
-        
-    ibias_err: float
-        The uncertainty in the bias current applied to the TES/shunt
-        system
-        
-    rsh: float
-        The shunt resistance in ohms
-        
-    rp: float
-        The parasitic resistance in ohms
-        
-    Returns
-    -------
-    bias_parameters_dict : dict
-        A dictonary of bias parameters and uncertainties, including i0, r0,
-        v0, and p0
-        
-    """
-
-    # check dimension
-    num_params = cov.shape[0]
-    if len(params.keys()) != num_params:
-        raise ValueError('ERROR: inconsistent number of '
-                         'parameters with covariance '
-                         'matrix shape')
-    
-    # Rload
-    rl = rp + rsh
-
-    # r0
-    dvdi0 = None
-    r0_jac = None
-    if num_params == 5:
-        dvdi0 =  params['A'] +  params['B']
-        r0_jac = np.asarray([1, 1, 0, 0, 0])
-    else:
-        dvdi0 =  params['A'] + params['B']/(1-params['C'])
-        r0_jac = np.asarray([1, 1, 1, 0, 0, 0, 0])
-        
-    r0 = abs(dvdi0) + rl
-    r0_err = np.matmul(r0_jac, np.matmul(cov, np.transpose(r0_jac)))
-    
-    i0 = ibias * rsh / (r0 + rl)
-    i0_err = ((ibias_err * rsh / (r0 + rl))**2 + (r0_err * ibias * rsh * (rl + r0)**-2)**2)**0.5
-    
-    
-    v0, v0_err = _get_v0(i0, i0_err, ibias, ibias_err, rsh, rp)
-    p0, p0_err = _get_p0(i0, i0_err, v0, v0_err)
-    
-    if rn is None:
-        rn = np.nan
- 
-    bias_parameter_dict = {
-        'i0': i0,
-        'i0_err': i0_err,
-        'v0': v0,
-        'v0_err': v0_err,
-        'r0': r0,
-        'r0_err': r0_err,
-        'p0': p0,
-        'p0_err': p0_err,
-        'rp': rp,
-        'rsh': rsh,
-        'rshunt': rsh,
-        'rl': rl,
-        'ibias': ibias,
-        'rn': rn,
-    }
-    
-    return bias_parameter_dict
+    response = _BaseDIDV._convolvedidv(
+                                       t, A, B, C,
+                                       tau1, tau2, tau3,
+                                       sgamp, rsh, sgfreq, dutycycle,
+                                       )
+                                   
+    return response
