+++ qetpy/core/didv/_didv.py
@@ -1,9 +1,9 @@
 import numpy as np
 from scipy.optimize import least_squares, fsolve
-from ._base_didv import _BaseDIDV, complexadmittance, get_i0, get_ibias
-from ._base_didv import get_tes_bias_parameters_dict, get_tes_bias_parameters_dict_infinite_loop_gain
+from ._base_didv import _BaseDIDV, complexadmittance
+from qetpy.core._biasparams import get_biasparams_offsets, get_biasparams_ilg
 from ._plot_didv import _PlotDIDV
-from ._uncertainties_didv import get_smallsignalparams_cov, get_smallsignalparams_sigmas
+from ._uncertainties_didv import get_smallsignalparams_vals, get_smallsignalparams_cov, get_smallsignalparams_sigmas
 from qetpy.utils import fft, ifft, fftfreq, rfftfreq
 import copy
 import warnings
@@ -460,9 +460,7 @@ class DIDV(_BaseDIDV, _PlotDIDV):
               guess_isloopgainsub1=None,
               lgcfix=None, verbose=0, max_nfev=1000,
               method='trf', loss='linear',
-              ftol=1e-15, xtol=1e-15,
-              biasparams_dict=None,
-              lgc_ssp_light=False):
+              ftol=1e-15, xtol=1e-15,):
         """
         This method does the fit that is specified by the variable
         poles. If the `processtraces` method has not been run yet, then
@@ -477,10 +475,6 @@ class DIDV(_BaseDIDV, _PlotDIDV):
             the specified fitting routine. Default is `np.inf`, which
             is equivalent to no cutoff frequency.
         bounds: 
-        lgc_ssp_light : bool, optional (3-poles only)
-            Used to tell dofit that the smallsignalparams light (only
-            beta, l, L, tau0, gratio) result dictionary including
-            uncertainties and covaraiance matrix should be calculted
 
         Raises
         ------
@@ -584,10 +578,6 @@ class DIDV(_BaseDIDV, _PlotDIDV):
             # store offset 
             self._fit_results[1]['offset'] = self._offset
             self._fit_results[1]['offset_err'] = self._offset_err
-
-            # calculate small signal parameters
-            # (lgc_ssp_light only used for 3-poles)
-            self._calc_ssp(1)
             
             
 
@@ -696,12 +686,6 @@ class DIDV(_BaseDIDV, _PlotDIDV):
             # store a few more parameters
             self._fit_results[2]['offset'] = self._offset
             self._fit_results[2]['offset_err'] = self._offset_err
-
-            # calculate small signal parameter
-            # (lgc_ssp_light only used for 3-poles)
-            self._calc_ssp(2,
-                           biasparams_dict=biasparams_dict,
-                           lgc_ssp_light=lgc_ssp_light)
 
         elif poles==3:
 
@@ -813,25 +797,14 @@ class DIDV(_BaseDIDV, _PlotDIDV):
             self._fit_results[3]['offset'] = self._offset
             self._fit_results[3]['offset_err'] = self._offset_err
             
-            # small signal parameters
-            self._calc_ssp(
-                3,
-                biasparams_dict=biasparams_dict,
-                lgc_ssp_light=lgc_ssp_light
-            )
 
         else:
             raise ValueError("The number of poles should be 1, 2, or 3.")
 
 
-    def calc_smallsignal_params(self, ivsweep_results,
+    def calc_smallsignal_params(self, biasparams,
+                                biasparams_tag = 'normal_iv',
                                 poles=None,
-                                calc_true_current=False,
-                                tes_bias=None,
-                                close_loop_norm=None,
-                                output_variable_gain=None,
-                                output_variable_offset=None,
-                                inf_loop_gain_approx='auto',
                                 lgc_verbose=True,
                                 lgc_diagnostics=False):
         """
@@ -870,81 +843,26 @@ class DIDV(_BaseDIDV, _PlotDIDV):
         # True bias calculation
         
         # check ivsweep results and convert to dictionary
-        if not isinstance(ivsweep_results, dict):
-            raise ValueError(f'ERROR: "ivsweep_results" should be a dictionary!')
+        if not isinstance(biasparams, dict):
+            raise ValueError(f'ERROR: "biasparams" should be a dictionary!')
         
-        required_parameters = ['rp']
-        if calc_true_current:
-            required_parameters.extend(['i0_off', 'i0_off_err',
-                                        'ibias_off', 'ibias_off_err'])
-        else:
-            required_parameters.extend(['i0', 'i0_err','r0', 'r0_err'])
-            
+        required_parameters = ['rp', 'i0', 'i0_err','r0', 'r0_err']
         for par in required_parameters:
-            if par not in ivsweep_results.keys():
+            if par not in biasparams.keys():
                 raise ValueError(f'ERROR: parameter {par} not found in '
                                  '"ivsweep_results" dictionary!')
- 
-        # check other parameterts if calc_true_current is True
-        if calc_true_current:
-
-            # variable offset (check other name for back compatibility)
-            if ('i0_variable_offset' not in ivsweep_results
-                and 'i0_changable_offset' not in ivsweep_results):
-                raise ValueError(f'ERROR: i0 variable offset not found in '
-                                 '"ivsweep_results" dictionary!')
-
-            # tes bias
-            if tes_bias is None:
-                raise ValueError(f'ERROR: "tes_bias" parameter (QET bias) '
-                                 'required when calculating true current')
 
         # initialize  bias parameters dict
-        biasparams_dict = ivsweep_results.copy()
-        biasparams_dict['biasparams_type'] = 'external_current'
+        biasparams_dict = biasparams.copy()
+        biasparams_dict['biasparams_type'] = biasparams_tag
         
-        rp = ivsweep_results['rp']
+        rp = biasparams_dict['rp']
         rn = None
-        if 'rn' in  ivsweep_results:
-            rn = ivsweep_results['rn']
+        if 'rn' in biasparams_dict:
+            rn = biasparams_dict['rn']
 
-        
-        # calculate true i0 and true tes bias
-        ibias = tes_bias
+        ibias = biasparams_dict['ibias']
         ibias_err = 0
-        
-        if calc_true_current:
-            
-            if lgc_verbose:
-                print(f'INFO: Calculating true current!')
-
-                      
-            # was offset inverted
-            lgc_invert_offset = False
-            if ('lgc_invert_offset' in ivsweep_results.keys()
-                and  ivsweep_results['lgc_invert_offset']):
-                lgc_invert_offset = True
-            
-            # calculate true i0
-            i0, i0_err = get_i0(self._offset, self._offset_err,
-                                ivsweep_results,
-                                output_variable_offset,
-                                close_loop_norm,
-                                output_variable_gain,
-                                lgc_invert_offset=lgc_invert_offset,
-                                lgc_diagnostics=lgc_diagnostics)
-
-            # calculate true ibias (QET bias)
-            ibias, ibias_err = get_ibias(tes_bias, ivsweep_results,
-                                         lgc_diagnostics=lgc_diagnostics)
-
-        
-            # recalculate v0, r0 with true current and store in dictionary
-            biasparams_dict = get_tes_bias_parameters_dict(
-                i0, i0_err, ibias, ibias_err, self._rsh, rp, rn=rn
-            )
-            
-            biasparams_dict['biasparams_type'] = 'true_current'
        
         self._r0 = biasparams_dict['r0']
         self._rp = biasparams_dict['rp']
@@ -961,270 +879,30 @@ class DIDV(_BaseDIDV, _PlotDIDV):
             # calc
             self._calc_ssp(model_poles,
                            biasparams_dict=biasparams_dict.copy(),
+                           biasparams_tag=biasparams_tag, 
                            lgc_ssp_light=True)
             
-            # Check if infinite loop gain needs to be done
-            # if inf_loop_gain_approx == 'auto' AND lopp gain negative
-            #     -> automatically make approximation that loop gain infinite 
-
-            results = copy.deepcopy(self._fit_results[model_poles])
-                            
-            # check if infinite loop gain needs to be done
-            set_infinite_loop_gain = inf_loop_gain_approx
-            if inf_loop_gain_approx == 'auto':
                 
-                if results['smallsignalparams']['l'] < 0:
-                    if lgc_verbose:
-                        print(f'INFO: Loop gain is negative '
-                              f'for {model_poles}-poles fit. Will '
-                              f'use infinite loop gain approximation!')
-                        
-                    set_infinite_loop_gain = True
-                else:
-                    set_infinite_loop_gain = False
-                    
-            if set_infinite_loop_gain:
-
-                if lgc_verbose:
-                    print(f'INFO: Calculating bias parameters '
-                          f'with infinite loop gain approximation '
-                          f'for {model_poles}-poles fit!')
-
-                params =  results['params']
-                cov =  results['cov']
-                
-                biasparams_dict_inf = get_tes_bias_parameters_dict_infinite_loop_gain(
-                    params, cov, ibias, ibias_err, self._rsh, rp, rn=rn
-                )
-
-                biasparams_dict_inf['biasparams_type'] = 'infinite_lgain_current'
-                
-                # re-assign r0 
-                self._r0 = biasparams_dict_inf['r0']
-
-                # re-calculate small signal parameters
-                # with proper bias paramaters
-                self._calc_ssp(model_poles,
-                               biasparams_dict=biasparams_dict_inf,
-                               lgc_ssp_light=True)
-                
-        
-    def calc_bias_params_infinite_loop_gain(self, poles=None,
-                                            tes_bias=None,
-                                            tes_bias_err=0,
-                                            rp=None):
-        """
-        Calculate I0,R0,V0, P0 with infinite loop gain 
-        approximation and store in fit result
-        """
-
-        model_list  = list()
-
-        if poles is not None:
-            if (self._fit_results[poles] is None
-                or 'params' not in  self._fit_results[poles]):
-                raise ValueError(f'ERROR: {poles}-poles fit needs to be done first!')
-            else:
-                model_list.append(poles)
-        else:
-            if self._fit_results[2] is not None:
-                if  'params' not in  self._fit_results[2]:
-                    raise ValueError(f'ERROR: 2-poles fit needs to be done first!')
-                else:
-                    model_list.append(2)
-              
-            if self._fit_results[3] is not None:
-                if  'params' not in self._fit_results[3]:
-                    raise ValueError(f'ERROR: 3-poles fit needs to be done first!')
-                else:
-                    model_list.append(3)
-                    
-        if not model_list:
-            print('WARNING: No fit have been done. Doing nothing...')
-            return
-        
-        # loop poles
-        for model_poles in model_list:
-            
-            results = copy.deepcopy(self._fit_results[model_poles])
-            
-            # check if ibias available
-            if tes_bias is None:
-
-                if ('biasparams' not in results.keys()
-                    or 'ibias' not in results['biasparams']):
-                    raise ValueError(
-                        'ERROR: Unable to find tes bias (ibias)!'
-                        ' It needs to be provided.'
-                    )
-            
-                tes_bias = results['biasparams']['ibias']
-                if 'ibias_err' in results['biasparams']:
-                    tes_bias_err = results['biasparams']['ibias_err']
-          
-
-            # Rp
-            if rp is None:
-                if self._rp is None:
-                    raise ValueError(
-                        'ERROR: Unable to find rp!'
-                        ' It needs to be provided.'
-                    )
-                rp = self._rp
-
-            
-            results['biasparams_infinite_lgain'] = (
-                get_tes_bias_parameters_dict_infinite_loop_gain(
-                    results['params'], results['cov'],
-                    tes_bias, tes_bias_err,
-                    self._rsh, rp
-                )
-            )
-
-            self._fit_results[model_poles] = results
-                                  
-    def dofit_with_true_current(self, offset_dict,
-                                output_offset, closed_loop_norm, output_gain,
-                                ibias_metadata,
-                                bounds=None, guess=None,
-                                inf_loop_gain_approx=False,
-                                inf_loop_gain_limit=False,
-                                lgc_calibration_on=False,
-                                lgcdiagnostics=False):
-        """
-        Given the offset dictionary used to store the various current
-        current offsets used to reconstruct the true current through the 
-        TES and the trace metadata, finds the true current through the TES,
-        makes a biasparams dict, and recalculates the smallsignalparams
-        from the newly found true bias point.
-        
-        
-        Parameters:
-        ----------
-        
-        offset_dict: dict
-            Where are the relevant offsets are stored. Generated from the IV
-            sweep.
-            
-        output_offset: float, volts
-            The output offset gotten from the event metadata. In units of volts,
-            we correct for volts to amps conversion with the closed loop norm.
-            
-        closed_loop_norm: float, volts/amp=ohms
-            The constant from the metadata used to translate the voltage measured by
-            the DAQ into a current coming into the input coil of the SQUIDs. In units of
-            volts/amp = ohms.
-            
-        output_gain: float, dimensionless
-            The dimensionless gain used to convert the output offset in volts to the 
-            equivilant offset voltage measured by the DAQ
-            
-        ibias_metadata: float
-            The ibias gotten from the event metadata, i.e. without correcting for
-            the ibias offset calculated from the IV curve
-            
-        bounds: array, optional
-            Passed to dofit.
-            
-        guess: array, optional
-            Passed to dofit
-            
-        inf_loop_gain_approx : bool, optional
-            Defaults to False. If True, calculates the biasparameters and the
-            rest of the fits using the infinite loop gain approximation.
-            
-        inf_loop_gain_limit : bool, optional
-            Defaults to False. If True, calculates the biasparameters and the
-            rest of the fits using the infinite loop gain approximation only if
-            the fit loopgain is negative.
-
-        lgc_calibration_on : bool, optional
-            By default False (i.e. not using the calibration). If True, uses the calibration_dict
-            to more closely approximate how changing the output_offset changes the current
-            measured.
-        
-        Returns:
-        --------
-            
-        result3: fitresult_dict
-            3 pole fit result with biasparams calculated, and the smallsignalparams
-            correctly calculated from the r0 calculated for the biasparams
-        
-        """
-
-        warnings.warn(
-            'WARNING: "dofit_with_true_current(..)" function is deprecated. '
-            'Use "calc_smallsignal_params(.., calc_true_current=True,...)" '
-            'instead.')
-  
-        self._rp = offset_dict['rp']
-
-        rsh = self._rsh
-        rp = self._rp
-
-        offset = self._offset
-        offset_err = self._offset_err
-        
-        if 'calibration_dict' in offset_dict:
-            calibration_dict = offset_dict['calibration_dict']
-        else:
-            calibration_dict = None
-
-        i0, i0_err = get_i0(offset, offset_err, offset_dict, output_offset,
-                            closed_loop_norm, output_gain, lgc_diagnostics=lgcdiagnostics,
-                            lgc_calibration_on=lgc_calibration_on, calibration_dict=calibration_dict)
-        ibias, ibias_err = get_ibias(ibias_metadata, offset_dict, lgc_diagnostics=lgcdiagnostics)
-        biasparams_dict = get_tes_bias_parameters_dict(i0, i0_err, ibias, ibias_err, rsh, rp)
-        
-        if inf_loop_gain_approx:
-            biasparams_dict = get_tes_bias_parameters_dict_infinite_loop_gain( 
-                self._fit_results[3]['params'], self._fit_results[3]['cov'],
-                ibias, ibias_err, rsh, rp
-            )
-        
-        self._r0 = biasparams_dict['r0']
-
-        result3 = self.dofit(3, bounds=bounds, guess_params=guess,
-                             biasparams_dict=biasparams_dict,
-                             lgc_ssp_light=True)
-                             
-        if inf_loop_gain_limit:
-            if self._fit_results[3]['smallsignalparams']['l'] < 0:
-                biasparams_dict = get_tes_bias_parameters_dict_infinite_loop_gain( 
-                    self._fit_results[3]['params'], 
-                    self._fit_results[3]['cov'], i0, 
-                    i0_err, ibias, ibias_err, 
-                    rsh, rp)
-                
-                result3 = self.dofit(3, bounds=bounds,
-                                     guess_params=guess,
-                                     biasparams_dict=biasparams_dict,
-                                     lgc_ssp_light=True)
-                             
-        return result3
-    
- 
     @staticmethod
-    def _fitresult(poles, params, cov, falltimes, cost,                   
+    def _fitresult(poles, params, cov, falltimes, cost,
                    lgcfix=None):
-
         """
         Function for converting data from different fit results to a
         results dictionary.
-
         """
-
+        
         result = dict()
         result['lgcfix'] = lgcfix
         result['params_array'] = params
         
         # errors
         errors = np.diag(cov)**0.5
+        
         if lgcfix is not None:
             errors = np.zeros_like(lgcfix, dtype=float)
             np.place(errors, lgcfix, 0.0)
             np.place(errors, ~lgcfix,  np.diag(cov)**0.5)
-        
+            
         if poles == 1:
             result['params'] = {
                 'A': params[0],
@@ -1237,9 +915,9 @@ class DIDV(_BaseDIDV, _PlotDIDV):
                 'tau2': errors[1],
                 'dt': errors[2],
             }
-
+            
         elif (poles == 2 or poles == 3):
-
+        
             result['params'] = {
                 'A': params[0],
                 'B': params[1],
@@ -1249,9 +927,7 @@ class DIDV(_BaseDIDV, _PlotDIDV):
                 'tau3': params[5],
                 'dt': params[6],
             }
-            
             result['cov'] = cov
-            
             result['errors'] = {
                 'A': errors[0],
                 'B': errors[1],
@@ -1261,45 +937,40 @@ class DIDV(_BaseDIDV, _PlotDIDV):
                 'tau3': errors[5],
                 'dt': errors[6],
             }
-
+            
         # other params
         result['falltimes'] = falltimes
         result['cost'] = cost
-      
         return result
-
-    
+        
+        
     def _calc_ssp(self, poles,
                   biasparams_dict=None,
+                  biasparams_tag='norm_iv', 
                   lgc_ssp_light=False):
         """
         Function to calculate small signal parameters  from fit result
-        
         """
-
-
+        
         # check if r0/rp available in biasparams_dict
         if biasparams_dict is not None:
             if 'r0' in biasparams_dict:
                 self._r0 = biasparams_dict['r0']
             if 'rp' in biasparams_dict:
                 self._rp = biasparams_dict['rp']
-            
-
-        
-        # 1-poles fit 
+                
+        # 1-poles fit
         if poles == 1:
-            
             if self._fit_results[1] is None:
                 raise ValueError(
                     'ERROR: No 1-poles fit done! Unable to '
                     'calculate small signal parameters ')
-            
+                    
             smallsignalparams = DIDV._converttotesvalues(
                 self._fit_results[1]['params_array'],
                 self._rsh, self._r0, self._rp
             )
-
+            
             self._fit_results[1]['smallsignalparams'] = {
                 'rsh': smallsignalparams[0],
                 'rp': smallsignalparams[1],
@@ -1310,34 +981,32 @@ class DIDV(_BaseDIDV, _PlotDIDV):
             self._fit_results[1]['didv0'] = (
                 complexadmittance(0, **self._fit_results[1]['smallsignalparams']).real
             )
-
+            
             # store bias params
-            if  biasparams_dict is not None:  
+            if  biasparams_dict is not None:
                 self._fit_results[1]['biasparams'] = biasparams_dict.copy()
             else:
                 self._fit_results[1]['biasparams'] = None
-                        
-            
-        # 2-poles and 3-poles fit     
+                
+        # 2-poles and 3-poles fit
         if poles == 2 or poles == 3:
-
             # get result dictionary
             if (self._fit_results[poles] is None
                 or 'params_array' not in self._fit_results[poles]):
                 raise ValueError(
                     f'ERROR: No {poles}-poles fit done! Unable to '
                     f'calculate small signal parameters ')
-            
-            # results 
+                    
+            # results
             results = copy.deepcopy(self._fit_results[poles])
-
+            
+            results['biasparams_tag'] = biasparams_tag
             
             # convert fit parameterts to smallsignalparams
             smallsignalparams = DIDV._converttotesvalues(
                 results['params_array'],
                 self._rsh, self._r0, self._rp
             )
-
             results['smallsignalparams'] = {
                 'rsh': smallsignalparams[0],
                 'rp': smallsignalparams[1],
@@ -1350,48 +1019,35 @@ class DIDV(_BaseDIDV, _PlotDIDV):
                 'tau3': smallsignalparams[8],
                 'dt': smallsignalparams[9],
             }
-
-                                          
             results['didv0'] = (
                 complexadmittance(0, **results['smallsignalparams']).real
-            )              
-
+            )
+            
             # store bias params
-            if  biasparams_dict is not None:  
+            if  biasparams_dict is not None:
                 results['biasparams'] = biasparams_dict.copy()
             else:
                 results['biasparams'] = None
                 
             # calculate small signal parameters cov/sigmas
             if lgc_ssp_light:
-
                 if biasparams_dict is None:
                     raise ValueError(
                         'ERROR: "biasparams_dict" required when '
                         'lgc_ssp_light=True'
                     )
-
+                    
+                ssp_light_vals = get_smallsignalparams_vals(results)
                 ssp_light_cov = get_smallsignalparams_cov(results)
                 ssp_light_sigmas = get_smallsignalparams_sigmas(results)
-
-                # store
-                ssp_light_vals = {
-                    'beta': smallsignalparams[3],
-                    'l': smallsignalparams[4],
-                    'L': smallsignalparams[5],
-                    'tau0': smallsignalparams[6],
-                    'gratio': smallsignalparams[7],
-                }
                 
                 results['ssp_light'] = {
                     'vals': ssp_light_vals,
                     'cov': ssp_light_cov,
                     'sigmas': ssp_light_sigmas,
                 }
-
-
+                
             # store results internally
             self._fit_results[poles] = results
             
-                
-                    
+
