+++ qetpy/sim/_sim.py
@@ -439,44 +439,13 @@ class TESnoise:
         
         if freqs is None:
             freqs = self.freqs
-            
-        if set(freqs).issubset(self.squid_noise_current_freqs):
-            if self.lgc_diagnostics:
-                print("SQUID noise over-sampled")
-            #all noise freqs are in the squid noise frequencies
-            inds = np.zeros_like(freqs)
-            i = 0
-            while i < len(inds):
-                where_arr = np.where(self.squid_noise_current_freqs == freqs[i])
-                if len(where_arr) > 0:
-                    inds[i] = where_arr[0]
-                else:
-                    inds[i] = 0.0
-                i += 1
-
-            print(self.squid_noise_current[inds])
-
-        elif set(self.squid_noise_current_freqs).issubset(freqs):
-            if self.lgc_diagnostics:
-                print("SQUID noise under-sampled")
-            #we'll use a squid noise for multiple frequencies in freqs
-            squid_noise = np.zeros_like(freqs)
-
-            i = 0
-            j = 0
-            while i < len(squid_noise)/2:
-                while (self.squid_noise_current_freqs[j] < freqs[i]) & (j < len(self.squid_noise_current_freqs)/2):
-                    j += 1
-                
-                squid_noise[i] = self.squid_noise_current[j]
-                squid_noise[-i] = self.squid_noise_current[j]
-                i += 1
-            return squid_noise
         
-        else:
-            if self.lgc_diagnostics:
-                print("INTERPOLATION FAILED! SQUID noise frequencies and")
-                print("noise frequencies do not match.")
+        sorted_indices = np.argsort(self.squid_noise_current_freqs)
+        squid_noise = np.interp(freqs,
+                                self.squid_noise_current_freqs[sorted_indices],
+                                self.squid_noise_current[sorted_indices])
+        
+        return squid_noise
         
 
     def s_psquid(self, freqs=None):
